优化目录结构,修改README文件
删除了不必要的keymap目录
