+++ lua/plugins/tools/copilot.lua
@@ -18,4 +18,9 @@
 --   },
 -- }
 
+-- return {
+--   "github/copilot.vim",
+--   event = "BufEnter",
+-- }
+
 return {}
