+++ lua/utils/lsp.lua
@@ -1,10 +1,43 @@
 local M = {}
 
-local on_attach = require("keymap.on_attach").lspconfig
+local map = vim.keymap.set
+local conf = require("nvconfig").ui.lsp
 
--- export on_attach & capabilities
-M.on_attach = on_attach
--- disable semanticTokens
+M.on_attach = function(bufnr, client)
+  local function opts(desc)
+    return { buffer = bufnr, desc = "LSP " .. desc }
+  end
+
+  -- local function map(mode, l, r, desc)
+  --   vim.keymap.set(mode, l, r, { buffer = bufnr, desc = desc })
+  -- end
+
+  map("n", "gD", vim.lsp.buf.declaration, opts "Go to declaration")
+  map("n", "gd", vim.lsp.buf.definition, opts "Go to definition")
+  map("n", "gi", vim.lsp.buf.implementation, opts "Go to implementation")
+  map("n", "<leader>sh", vim.lsp.buf.signature_help, opts "Show signature help")
+  map("n", "<leader>wa", vim.lsp.buf.add_workspace_folder, opts "Add workspace folder")
+  map("n", "<leader>wr", vim.lsp.buf.remove_workspace_folder, opts "Remove workspace folder")
+
+  map("n", "<leader>wl", function()
+    print(vim.inspect(vim.lsp.buf.list_workspace_folders()))
+  end, opts "List workspace folders")
+
+  map("n", "<leader>D", vim.lsp.buf.type_definition, opts "Go to type definition")
+
+  map("n", "<leader>ra", function()
+    require "nvchad.lsp.renamer"()
+  end, opts "NvRenamer")
+
+  map({ "n", "v" }, "<leader>ca", vim.lsp.buf.code_action, opts "Code action")
+  map("n", "gr", vim.lsp.buf.references, opts "Show references")
+
+  -- setup signature popup
+  if conf.signature and client.server_capabilities.signatureHelpProvider then
+    require("nvchad.lsp.signature").setup(client, bufnr)
+  end
+end
+
 M.on_init = function(client, _)
   if client.supports_method "textDocument/semanticTokens" then
     client.server_capabilities.semanticTokensProvider = nil
@@ -31,6 +64,7 @@ M.capabilities.textDocument.completion.completionItem = {
   },
 }
 
+-- lspconfig default lsp config
 M.defaults = function()
   dofile(vim.g.base46_cache .. "lsp")
   require "nvchad.lsp"
